fix: update existing huddle 01 project
