+++ packages/app-store/apps.schemas.generated.ts
@@ -15,6 +15,7 @@ import { appDataSchema as googlecalendar_zod_ts } from "./googlecalendar/zod";
 import { appDataSchema as googlevideo_zod_ts } from "./googlevideo/zod";
 import { appDataSchema as gtm_zod_ts } from "./gtm/zod";
 import { appDataSchema as hubspot_zod_ts } from "./hubspot/zod";
+import { appDataSchema as huddle01video_zod_ts } from "./huddle01video/zod";
 import { appDataSchema as intercom_zod_ts } from "./intercom/zod";
 import { appDataSchema as jelly_zod_ts } from "./jelly/zod";
 import { appDataSchema as jitsivideo_zod_ts } from "./jitsivideo/zod";
@@ -62,6 +63,7 @@ export const appDataSchemas = {
   googlevideo: googlevideo_zod_ts,
   gtm: gtm_zod_ts,
   hubspot: hubspot_zod_ts,
+  huddle01video: huddle01video_zod_ts,
   intercom: intercom_zod_ts,
   jelly: jelly_zod_ts,
   jitsivideo: jitsivideo_zod_ts,
