--- packages/app-store/huddle-01/DESCRIPTION.md
@@ -1,8 +1,0 @@
----
-items:
-  - 1.jpeg
-  - 2.jpeg
-  - 3.jpeg
----
-
-{DESCRIPTION}
