--- packages/app-store/huddle-01/lib/index.ts
@@ -1,1 +1,0 @@
-export { default as VideoApiAdapter } from "./VideoApiAdapter";
