+++ packages/app-store/huddle01video/_metadata.ts
@@ -1,6 +1,6 @@
-import { randomString } from "@calcom/lib/random";
 import type { AppMeta } from "@calcom/types/App";
 
+import { LocationType } from "../locations";
 import _package from "./package.json";
 
 export const metadata = {
@@ -9,26 +9,27 @@ export const metadata = {
   installed: true,
   type: "huddle01_video",
   variant: "conferencing",
-  categories: ["conferencing"],
+  categories: ["video", "conferencing"],
   logo: "icon.svg",
-  publisher: "huddle01.com",
-  url: "https://huddle01.com",
-  category: "conferencing",
-  slug: "huddle01",
+  publisher: "Huddle01",
+  url: "https://www.huddle01.com/",
+  slug: "huddle-01",
   title: "Huddle01",
   isGlobal: false,
   email: "support@huddle01.com",
   appData: {
     location: {
+      type: "integrations:huddle01video",
+      label: "Huddle01",
       linkType: "dynamic",
-      type: "integrations:huddle01",
-      label: "Huddle01 Video",
     },
   },
-  key: { apikey: randomString(12) },
+  key: { apikey: "" },
   dirName: "huddle01video",
-  concurrentMeetings: true,
-  isOAuth: false,
 } as AppMeta;
 
-export default metadata;
+export const locationOption = {
+  value: LocationType.Conferencing,
+  label: "Huddle01 Video",
+  icon: "Link",
+};
