+++ packages/app-store/huddle01video/api/add.ts
@@ -1,54 +1,26 @@
-import type { NextApiRequest, NextApiResponse } from "next";
+import type { NextApiRequest } from "next";
+import { stringify } from "querystring";
 
-import { throwIfNotHaveAdminAccessToTeam } from "@calcom/app-store/_utils/throwIfNotHaveAdminAccessToTeam";
-import prisma from "@calcom/prisma";
+import { defaultHandler, defaultResponder } from "@calcom/lib/server";
 
-import getInstalledAppPath from "../../_utils/getInstalledAppPath";
+import { encodeOAuthState } from "../../_utils/oauth/encodeOAuthState";
 
-/**
- * This is an example endpoint for an app, these will run under `/api/integrations/[...args]`
- * @param req
- * @param res
- */
-export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  if (!req.session?.user?.id) {
-    return res.status(401).json({ message: "You must be logged in to do this" });
-  }
-  const { teamId, returnTo } = req.query;
+const WEBAPP_URL = "http://localhost:3000";
+async function handler(req: NextApiRequest) {
+  const state = encodeOAuthState(req);
 
-  await throwIfNotHaveAdminAccessToTeam({ teamId: Number(teamId) ?? null, userId: req.session.user.id });
-  const installForObject = teamId ? { teamId: Number(teamId) } : { userId: req.session.user.id };
+  const params = {
+    response_type: "code",
+    redirect_uri: `${WEBAPP_URL}/api/integrations/huddle-01/callback`,
+    state,
+    appName: "calcom",
+  };
+  const query = stringify(params);
+  const url = `https://darshan.huddle01.app/g_auth?${query}`;
 
-  const appType = "huddle01_video";
-  try {
-    const alreadyInstalled = await prisma.credential.findFirst({
-      where: {
-        type: appType,
-        ...installForObject,
-      },
-    });
-    if (alreadyInstalled) {
-      throw new Error("Already installed");
-    }
-    const installation = await prisma.credential.create({
-      data: {
-        type: appType,
-        key: {},
-        ...installForObject,
-        appId: "huddle01",
-      },
-    });
-    if (!installation) {
-      throw new Error("Unable to create user credential for huddle01video");
-    }
-  } catch (error: unknown) {
-    if (error instanceof Error) {
-      return res.status(500).json({ message: error.message });
-    }
-    return res.status(500);
-  }
-  // need to return a json object with the response status
-  return res
-    .status(200)
-    .json({ url: returnTo ?? getInstalledAppPath({ variant: "conferencing", slug: "huddle01" }) });
+  return { url };
 }
+
+export default defaultHandler({
+  GET: Promise.resolve({ default: defaultResponder(handler) }),
+});
