+++ packages/app-store/huddle01video/api/index.ts
@@ -1,1 +1,2 @@
 export { default as add } from "./add";
+export { default as callback } from "./callback";
