+++ packages/app-store/huddle01video/index.ts
@@ -1,3 +1,3 @@
-export * as lib from "./lib";
 export * as api from "./api";
+export * as lib from "./lib";
 export { metadata } from "./_metadata";
