+++ packages/app-store/huddle01video/zod.ts
@@ -1,0 +1,7 @@
+import { z } from "zod";
+
+export const appDataSchema = z.object({});
+
+export const appKeysSchema = z.object({
+  api: z.string().min(1),
+});
